Fixed compiler errors.

+++ Proxy/Proxy/Listener.cs
@@ -199,8 +199,8 @@ public abstract class Listener : IDisposable{
 	///<param name="IP">The IP address to check.</param>
 	///<returns>True if the specified IP address is a remote address, false otherwise.</returns>
 	protected static bool IsRemoteIP(IPAddress IP) {
-		byte First = (byte)Math.Floor(IP.Address % 256);
-		byte Second = (byte)Math.Floor((IP.Address % 65536) / 256);
+		byte First = (byte)Math.Floor((double)(IP.Address % 256));
+		byte Second = (byte)Math.Floor((double)((IP.Address % 65536) / 256));
 		//Not 10.x.x.x And Not 172.16.x.x <-> 172.31.x.x And Not 192.168.x.x
 		//And Not Any And Not Loopback And Not Broadcast
 		return (First != 10) &&
@@ -214,8 +214,8 @@ public abstract class Listener : IDisposable{
 	///<param name="IP">The IP address to check.</param>
 	///<returns>True if the specified IP address is a local address, false otherwise.</returns>
 	protected static bool IsLocalIP(IPAddress IP) {
-		byte First = (byte)Math.Floor(IP.Address % 256);
-		byte Second = (byte)Math.Floor((IP.Address % 65536) / 256);
+		byte First = (byte)Math.Floor((double)(IP.Address % 256));
+		byte Second = (byte)Math.Floor((double)((IP.Address % 65536) / 256));
 		//10.x.x.x Or 172.16.x.x <-> 172.31.x.x Or 192.168.x.x
 		return (First == 10) ||
 			(First == 172 && (Second >= 16 && Second <= 31)) ||
